Refactoring: Separation of Concerns
